requirements.txt
streamlit == 0.75.0
pandas == pandas
plotly == 4.14.1
sqlalchemy == 1.3.18
python-dotenv == 0.15.0
mysql-connector-python == 8.0.23